requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
python-dotenv>=1.0
scikit-learn>=1.4
# For live trading (optional)
MetaTrader5>=5.0; platform_system == "Windows"
